+++ java-io-http/src/main/java/nbbrd/io/http/HttpAuthenticator.java
@@ -1,24 +1,24 @@
 package nbbrd.io.http;
 
+import internal.io.http.NoOpHttpAuthenticator;
 import lombok.NonNull;
 import nbbrd.design.StaticFactoryMethod;
 import org.jspecify.annotations.Nullable;
 
 import java.io.IOException;
 import java.net.PasswordAuthentication;
-import java.net.URL;
+import java.net.URI;
 
 public interface HttpAuthenticator {
 
     @Nullable
-    PasswordAuthentication getPasswordAuthentication(@NonNull URL url) throws IOException;
+    PasswordAuthentication getPasswordAuthentication(@NonNull URI uri) throws IOException;
 
-    default void invalidate(@NonNull URL url) throws IOException {
-    }
+    void invalidate(@NonNull URI uri) throws IOException;
 
     @StaticFactoryMethod
     static @NonNull HttpAuthenticator noOp() {
-        return ignore -> null;
+        return NoOpHttpAuthenticator.NO_OP;
     }
 
     @StaticFactoryMethod(PasswordAuthentication.class)