+++ java-io-http/src/main/java/nbbrd/io/http/UrlConnectionListener.java
@@ -7,29 +7,19 @@ import nbbrd.design.NotThreadSafe;
 import nbbrd.design.StaticFactoryMethod;
 
 import java.net.Proxy;
-import java.net.URL;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
 @NotThreadSafe
 public interface UrlConnectionListener {
 
-    default void onOpen(@NonNull HttpRequest request, @NonNull Proxy proxy, @NonNull HttpAuthScheme scheme) {
+    default void onOpen(@NonNull HttpRequest request, @NonNull Proxy proxy) {
     }
 
     default void onSuccess(@NonNull Supplier<String> contentType) {
     }
 
-    default void onRedirection(@NonNull URL oldUrl, @NonNull URL newUrl) {
-    }
-
-    default void onUnauthorized(@NonNull URL url, @NonNull HttpAuthScheme oldScheme, @NonNull HttpAuthScheme newScheme) {
-    }
-
     default void onEvent(@NonNull String message) {
-    }
-
-    default void onComplete(@NonNull HttpRequest request, long bytesRead, long elapsedMs) {
     }
 
     @StaticFactoryMethod