+++ java-io-http/src/main/java/nbbrd/io/http/ext/RedirectListener.java
@@ -1,0 +1,18 @@
+package nbbrd.io.http.ext;
+
+import lombok.NonNull;
+import nbbrd.design.StaticFactoryMethod;
+
+import java.net.URI;
+
+@FunctionalInterface
+public interface RedirectListener {
+
+    void onRedirection(@NonNull URI oldUri, @NonNull URI newUri);
+
+    @StaticFactoryMethod
+    static @NonNull RedirectListener noOp() {
+        return (oldUri, newUri) -> {
+        };
+    }
+}