+++ CHANGELOG.md
@@ -16,6 +16,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Refactor HTTP response exception handling as a decorator
+- Move UrlConnectionHttpClient features to decorators
 
 ## [0.0.38] - 2026-06-02